Take the JVM and the neutral form out of the example verifier
Three responsibilities were in one mutable object: what an example means, what a
value looks like in the form a decoder reads, and how a helper is run on the
JVM. Issue #200 added to all three at once, which is why the row's timeout came
to share state with the rows after it.

`HelperInvoker` runs a helper as the static method its module emits: the class,
the method name, the reflection, and what a failure inside the helper means.
The fixture that applies a helper hands over a name and its arguments and reads
back a value or the reason the row can state nothing.

`NeutralForm` says what the neutral form of a value is. A row reaches that form
two ways — from what the author wrote, and from what a helper returned — and
the rules that decide the form itself (which discriminator a case carries, when
a unit case is a bare name, which written list is a map's entries) are the same
rules either way, so they are in one place rather than restated on one side.
Re-materialising a live value is its entry point for the second direction.

`FixtureException` and `NonTerminationException` become their own files: they
are now the vocabulary three classes share rather than one class's internals.

ExampleVerifier keeps what an example means: the target, the row's checks, the
fakes, the comparison and the diagnostics. It is 1392 lines from 1700, and
smaller than before issue #200 was implemented.
