+++ souther-compiler/src/main/java/souther/compiler/ExampleVerifier.java
@@ -43,7 +43,13 @@ import java.util.Set;
  * {@code with dep = value} on the row (a constant/value dependency) or a {@code fake dep | table}
  * declaration (an input-keyed function dependency). Each fake is a {@code Behavior} proxy passed to
  * the generated behavior's injecting constructor — it produces no run-time class. Inputs, expected
- * values, and fake entries are fixtures — literals, newtype constructions, and record constructions.
+ * values, and fake entries are fixtures — literals, newtype constructions, record constructions, and
+ * helpers applied to those (ADR-0077).
+ *
+ * <p>Three things a row needs are not here, because they are not about what an example means:
+ * {@link NeutralForm} says what a value looks like in the form a decoder reads, {@link HelperInvoker}
+ * runs a helper as the method its module emits, and {@link RowEvaluation} owns the state one row builds
+ * up while it is evaluated.
  */
 public final class ExampleVerifier {
 
@@ -143,6 +149,10 @@ public final class ExampleVerifier {
     private final MemoryClassLoader loader;
     /** The values a row may name: this module's own, and the ones its imports bring in. */
     private final Map<String, Ast.FnDef> values;
+    /** Runs a helper a fixture applies. One evaluation's, because what it is running is one row's. */
+    private final HelperInvoker helpers;
+    /** What a value looks like in the form a decoder reads — the rules both directions of a row read. */
+    private final NeutralForm neutral;
 
     private ExampleVerifier(Ast.Module module, Symbols symbols, Map<String, Sig> sigs,
                             MemoryClassLoader loader, Map<String, Ast.FnDef> values) {
@@ -151,6 +161,8 @@ public final class ExampleVerifier {
         this.sigs = sigs;
         this.loader = loader;
         this.values = values;
+        this.helpers = new HelperInvoker(module.name(), loader);
+        this.neutral = new NeutralForm(symbols);
     }
 
     // --- one example (a target and its rows) --------------------------------------------------
@@ -269,7 +281,7 @@ public final class ExampleVerifier {
 
         /** The helper this row is inside, for the budget to name when the row does not finish. */
         String helper() {
-            return evaluation.applying.get();
+            return evaluation.helpers.running();
         }
     }
 
@@ -647,7 +659,7 @@ public final class ExampleVerifier {
      * (built before the behavior ran), by structural equality. */
     private boolean matches(Ast.Expr expected, Object result, Object expectedValue) {
         if (expected instanceof Ast.Var v) {
-            return simpleName(result).equals(v.name());
+            return NeutralForm.simpleName(result).equals(v.name());
         }
         return expectedValue != null && expectedValue.equals(result);
     }
@@ -661,7 +673,7 @@ public final class ExampleVerifier {
             return decode(Type.ref(nd.typeName().denotes()),
                     raw(expected, Type.ref(nd.typeName().denotes())));
         }
-        if (expected instanceof Ast.Call c && isNewtype(c.fn())) {
+        if (expected instanceof Ast.Call c && neutral.isNewtype(c.fn())) {
             return decode(Type.ref(symbols.resolve(c.fn())), raw(expected));
         }
         // A collection output has no case name to decode against, so the behavior's declared
@@ -695,7 +707,7 @@ public final class ExampleVerifier {
         if (expected instanceof Ast.NewData nd) {
             return nd.typeName().written();
         }
-        if (expected instanceof Ast.Call c && isNewtype(c.fn())) {
+        if (expected instanceof Ast.Call c && neutral.isNewtype(c.fn())) {
             return c.fn();
         }
         return null;   // a literal expected (a primitive output)
@@ -741,7 +753,7 @@ public final class ExampleVerifier {
      * writes, so the two sides of a mismatch can be read against each other. A value with no encoder
      * (or one that fails to encode) falls back to its case name alone. */
     private String describeActual(Object result) {
-        String name = simpleName(result);
+        String name = NeutralForm.simpleName(result);
         if (name.isEmpty()) {
             return String.valueOf(result);
         }
@@ -752,7 +764,7 @@ public final class ExampleVerifier {
         if (encoded != null) {
             return show(name, encoded);
         }
-        return isScalar(result) ? showValue(result) : name;
+        return NeutralForm.isScalar(result) ? showValue(result) : name;
     }
 
     /** A live value in the notation a fixture is written in, at any depth: a collection element by
@@ -760,7 +772,7 @@ public final class ExampleVerifier {
      * mismatch go through this, so they can be read against each other — and neither shows the JDK
      * class that happened to carry the collection. */
     private String showAny(Object v) {
-        if (v == null || isScalar(v)) {
+        if (v == null || NeutralForm.isScalar(v)) {
             return showValue(v);
         }
         if (v instanceof Map<?, ?> m) {
@@ -777,16 +789,9 @@ public final class ExampleVerifier {
             }
             return elements.isEmpty() ? "[]" : "[ " + String.join(", ", elements) + " ]";
         }
-        String name = simpleName(v);
+        String name = NeutralForm.simpleName(v);
         Object encoded = encodedOrNull(v, name);
         return encoded != null ? show(name, encoded) : name;
-    }
-
-    /** Whether a value is a neutral scalar, so it can be shown as written rather than by class name. */
-    private static boolean isScalar(Object v) {
-        return v instanceof String || v instanceof Long || v instanceof Boolean
-                || v instanceof BigDecimal || v instanceof java.time.LocalDate
-                || v instanceof java.time.LocalDateTime;
     }
 
     /**
@@ -874,15 +879,6 @@ public final class ExampleVerifier {
         }
     }
 
-    private static String simpleName(Object o) {
-        if (o == null) {
-            return "null";
-        }
-        String n = o.getClass().getName();
-        int dot = n.lastIndexOf('.');
-        return dot < 0 ? n : n.substring(dot + 1);
-    }
-
     private Set<TypeName> outCases(Type out) {
         if (out instanceof Type.Union || out instanceof Type.Ref) {
             return TypeOps.leafCases(out, symbols);
@@ -917,7 +913,7 @@ public final class ExampleVerifier {
             case Ast.Var v -> unitInput(v.name(), expected);
             case Ast.NewData nd -> record(nd);
             case Ast.ListLit l -> {
-                Type element = elementOf(expected);
+                Type element = NeutralForm.elementOf(expected);
                 List<Object> out = new ArrayList<>();
                 for (Ast.Expr el : l.elements()) {
                     out.add(raw(el, element));
@@ -927,7 +923,7 @@ public final class ExampleVerifier {
             // a `(key, value)` pair: only a `Map` field's entries are written this way, and `shaped`
             // collects them into the map the decoder reads (a tuple is not a data field type itself).
             case Ast.Tuple t -> {
-                List<Type> parts = entryTypes(expected, t.elements().size());
+                List<Type> parts = NeutralForm.entryTypes(expected, t.elements().size());
                 List<Object> out = new ArrayList<>();
                 for (int i = 0; i < t.elements().size(); i++) {
                     out.add(raw(t.elements().get(i), parts == null ? null : parts.get(i)));
@@ -936,34 +932,6 @@ public final class ExampleVerifier {
             }
             default -> throw new FixtureException("an example fixture must be a literal or a construction");
         };
-    }
-
-    /** What a written list holds, when the position says: a list's or a set's element, or a map's
-     * entry pair. An optional is opened first — writing a value for a `T?` field writes a `T`. */
-    private static Type elementOf(Type expected) {
-        return switch (open(expected)) {
-            case Type.ListOf l -> l.element();
-            case Type.SetOf s -> s.element();
-            case Type.MapOf m -> Type.tuple(List.of(m.key(), m.value()));
-            case null, default -> null;
-        };
-    }
-
-    /** The types of a written tuple's parts, or null when the position does not say. A map's entry is
-     * the pair that carries its key type, which is where an enumeration key is written. */
-    private static List<Type> entryTypes(Type expected, int arity) {
-        Type opened = open(expected);
-        if (opened instanceof Type.MapOf m && arity == 2) {
-            return List.of(m.key(), m.value());
-        }
-        if (opened instanceof Type.TupleOf t && t.elements().size() == arity) {
-            return t.elements();
-        }
-        return null;
-    }
-
-    private static Type open(Type t) {
-        return t instanceof Type.OptionOf o ? open(o.element()) : t;
     }
 
     /** A bare name in a fixture is only meaningful as a unit case (no fields) or the built-in {@code
@@ -981,11 +949,11 @@ public final class ExampleVerifier {
             // be a case of an enumeration and of a sum that has a field-bearing case, and those travel
             // differently — so the position's own type decides, and the case's sums answer only where
             // the position does not say.
-            if (caseName != null && readsABareName(expected, caseName)) {
+            if (caseName != null && neutral.readsABareName(expected, caseName)) {
                 return caseName.name();
             }
             Map<String, Object> unit = new LinkedHashMap<>();
-            tagged(name, unit);   // a unit case of a sum still needs the tag its decoder reads
+            neutral.tagged(name, unit);   // a unit case of a sum still needs the tag its decoder reads
             return unit;
         }
         Ast.Expr value = valueBody(name);
@@ -1032,35 +1000,6 @@ public final class ExampleVerifier {
         }
     }
 
-    /** Whether the position this case is written in reads a bare name: it is typed as an enumeration,
-     * or it is untyped here and every sum that lists the case is one. */
-    private boolean readsABareName(Type expected, TypeName caseName) {
-        Type position = open(expected);
-        return position != null
-                ? TypeOps.isUnitOnlySum(position, symbols)
-                : onlyEnumerationsList(caseName);
-    }
-
-    /** Whether every sum that lists this case is an enumeration, so its neutral form is its name
-     * wherever it is written. Asked only where the position has no declared type to read it as. */
-    private boolean onlyEnumerationsList(TypeName caseName) {
-        boolean listed = false;
-        for (Ast.Def def : symbols.visible()) {
-            if (!(def instanceof Ast.SumData sum) || sum.decoder().isEmpty()) {
-                continue;
-            }
-            for (Ast.Variant variant : sum.decoder().get().variants()) {
-                if (caseName.equals(variant.caseType().denotes())) {
-                    if (!TypeOps.isUnitOnlySum(sum, symbols)) {
-                        return false;
-                    }
-                    listed = true;
-                }
-            }
-        }
-        return listed;
-    }
-
     /**
      * {@code Set.fromList([…])} and {@code Map.fromList([…])} are the forms a fixture's own notation
      * stands for — a set is written as its elements and a map as its entry pairs — so the neutral
@@ -1075,7 +1014,7 @@ public final class ExampleVerifier {
             }
             return raw(c.args().get(0), expected);
         }
-        if (!isNewtype(c.fn()) && helperDef(c.fn()) instanceof Ast.FnDef helper) {
+        if (!neutral.isNewtype(c.fn()) && helperDef(c.fn()) instanceof Ast.FnDef helper) {
             return applied(c, helper, expected);
         }
         return newtypeInner(c);
@@ -1129,136 +1068,7 @@ public final class ExampleVerifier {
             Type paramType = TypeOps.resolveParamType(p.type(), symbols);
             args[i] = decode(paramType, raw(c.args().get(i), paramType));
         }
-        return neutral(run(c.fn(), args), expected, c.fn());
-    }
-
-    /** The helper currently being run, for the row's budget to name when it does not finish. One row's,
-     * because an evaluation is one row's ({@link RowEvaluation}); read from the thread waiting on that
-     * row, which is why it is not a plain field. */
-    private final java.util.concurrent.atomic.AtomicReference<String> applying =
-            new java.util.concurrent.atomic.AtomicReference<>();
-
-    /**
-     * Runs the helper's emitted method ({@code $Fns}) on the decoded arguments. A helper this module
-     * only expands has no method, which is what an intrinsic implemented in Java and one whose body
-     * produces a function are — a row applying either is refused by that reason rather than by the
-     * general rule about what a fixture may apply.
-     */
-    private Object run(String name, Object[] args) {
-        java.lang.reflect.Method method;
-        try {
-            Class<?> fns = loader.loadClass(module.name() + ".$Fns");
-            Class<?>[] params = new Class<?>[args.length];
-            java.util.Arrays.fill(params, Object.class);
-            method = fns.getDeclaredMethod(Backend.helperMethod(name), params);
-            method.setAccessible(true);
-        } catch (ClassNotFoundException | NoSuchMethodException _) {
-            throw new FixtureException("`" + name + "` cannot be called from an example fixture:"
-                    + " it has no executable helper method");
-        }
-        String outer = applying.get();
-        applying.set(name);
-        try {
-            return method.invoke(null, args);
-        } catch (java.lang.reflect.InvocationTargetException ite) {
-            Throwable cause = ite.getCause();
-            if (cause instanceof StackOverflowError) {
-                throw new NonTerminationException("`" + name + "` overflowed the stack");
-            }
-            throw new FixtureException("`" + name + "` did not produce a value: "
-                    + (cause == null ? ite : cause.getMessage()));
-        } catch (ReflectiveOperationException e) {
-            throw new FixtureException("`" + name + "` could not be called: " + e.getMessage());
-        } finally {
-            applying.set(outer);
-        }
-    }
-
-    /**
-     * A value a helper returned, in the neutral form a fixture is written in.
-     *
-     * <p>Directed by what the value is and by the position it lands in, not by encoding everything: a
-     * scalar and a temporal are already their own neutral form (a `Date` field's is a {@code LocalDate}
-     * while its encoder writes ISO text), a collection is re-materialised element by element, and a
-     * data goes through the derived encoder a mismatch is displayed through — with the discriminator
-     * its sum's decoder reads, as a written record fixture takes. A unit case of an enumeration travels
-     * as its name where the position reads one, which is the same question {@link #unitInput} asks.
-     */
-    private Object neutral(Object live, Type position, String helper) {
-        if (live == null || isScalar(live)) {
-            return live;
-        }
-        Type opened = open(position);
-        if (live instanceof Map<?, ?> m) {
-            Type key = opened instanceof Type.MapOf map ? map.key() : null;
-            Type value = opened instanceof Type.MapOf map ? map.value() : null;
-            Map<Object, Object> out = new LinkedHashMap<>();
-            for (Map.Entry<?, ?> e : m.entrySet()) {
-                out.put(neutral(e.getKey(), key, helper), neutral(e.getValue(), value, helper));
-            }
-            return out;
-        }
-        if (live instanceof Iterable<?> it) {
-            Type element = elementOf(opened);
-            List<Object> out = new ArrayList<>();
-            for (Object e : it) {
-                out.add(neutral(e, element, helper));
-            }
-            return out;
-        }
-        String name = simpleName(live);
-        // An optional field holds an `Option`: `None` is the absent value a fixture writes by leaving
-        // the field out, and `Some` holds what the position's own type reads (spec 8).
-        if (name.equals("Option$None")) {
-            return null;
-        }
-        if (name.equals("Option$Some")) {
-            return neutral(field(live, "value", helper), opened, helper);
-        }
-        TypeName caseName = symbols.resolve(name);
-        if (caseName == null) {
-            throw new FixtureException("`" + helper + "` returned a " + name
-                    + ", which is not a type this example can read");
-        }
-        if (!(symbols.get(caseName) instanceof Ast.Data data)) {
-            // a unit case: its name where the position reads one, else the tag its sum's decoder reads
-            if (readsABareName(position, caseName)) {
-                return caseName.name();
-            }
-            Map<String, Object> unit = new LinkedHashMap<>();
-            tagged(caseName, unit);
-            return unit;
-        }
-        if (data.newtype()) {
-            Type base = shapeOf(newtypeBaseType(caseName));
-            return adjacentlyTagged(name, shaped(neutral(field(live, "value", helper), base, helper),
-                    base));
-        }
-        Map<String, Ast.TypeRef> declared = fieldTypes(caseName);
-        Map<String, Object> out = new LinkedHashMap<>();
-        for (Map.Entry<String, Ast.TypeRef> f : declared.entrySet()) {
-            Type type = shapeOf(f.getValue());
-            Object value = shaped(neutral(field(live, f.getKey(), helper), type, helper), type);
-            // an absent optional is left out, the same neutral form a fixture writes for `None`
-            if (value != null) {
-                out.put(f.getKey(), value);
-            }
-        }
-        tagged(caseName, out);
-        return out;
-    }
-
-    /** One field of a live data, read through the accessor every data has (ADR-0065). Its class may be
-     * package-private, so the declared method is taken and opened, as a codec is. */
-    private Object field(Object live, String name, String helper) {
-        try {
-            java.lang.reflect.Method accessor = live.getClass().getDeclaredMethod(name);
-            accessor.setAccessible(true);
-            return accessor.invoke(live);
-        } catch (ReflectiveOperationException _) {
-            throw new FixtureException("the value `" + helper + "` returned cannot be read back as a"
-                    + " fixture: `" + simpleName(live) + "` has no `" + name + "` to read");
-        }
+        return neutral.of(helpers.invoke(c.fn(), args), expected, c.fn());
     }
 
     private Object newtypeInner(Ast.Call c) {
@@ -1270,7 +1080,7 @@ public final class ExampleVerifier {
             }
             return CallElaborator.parseTemporal(c.fn(), lit.value(), c.pos());
         }
-        if (!isNewtype(c.fn())) {
+        if (!neutral.isNewtype(c.fn())) {
             if (noMethod(c.fn())) {
                 // A function this module cannot run: an intrinsic implemented in Java, or a helper
                 // whose body produces a function. Said as that, so the rule that a fixture may apply a
@@ -1285,7 +1095,7 @@ public final class ExampleVerifier {
         }
         // a newtype over a temporal (`data 貸出日 = Date`) wraps the parsed value, so its written
         // string is read here as it would be for a bare `Date("...")`
-        String base = newtypeBase(c.fn());
+        String base = neutral.newtypeBase(c.fn());
         if (("Date".equals(base) || "DateTime".equals(base))
                 && c.args().get(0) instanceof Ast.StringLit lit) {
             return CallElaborator.parseTemporal(base, lit.value(), c.pos());
@@ -1293,26 +1103,9 @@ public final class ExampleVerifier {
         // the argument is shaped against what the newtype wraps, the same way a record fixture
         // shapes a field's value: a `Map` newtype's entry pairs become a map, a `Set` newtype's
         // written list stays a list for its decoder to dedupe
-        return adjacentlyTagged(c.fn(),
-                shaped(raw(c.args().get(0), shapeOf(newtypeBaseType(c.fn()))),
-                        shapeOf(newtypeBaseType(c.fn()))));
-    }
-
-    /**
-     * A newtype case of a sum decodes from the adjacent form its sum's decoder reads — the inner value
-     * under {@code value}, next to the discriminator (spec 10.3) — while a fixture names the case the
-     * way the domain constructs it, {@code アクティベート済み(メールアドレス("a@example.com"))}. Wrap it
-     * here, as {@link #record} does for a product case and {@link #unitInput} for a unit one; a newtype
-     * that is nobody's case is its bare inner value, unchanged.
-     */
-    private Object adjacentlyTagged(String caseName, Object inner) {
-        Map<String, Object> tagged = new LinkedHashMap<>();
-        tagged(caseName, tagged);
-        if (tagged.isEmpty()) {
-            return inner;   // not a case of any sum: the newtype's own form is its inner value
-        }
-        tagged.put("value", inner);
-        return tagged;
+        return neutral.adjacentlyTagged(c.fn(),
+                neutral.shaped(raw(c.args().get(0), neutral.shapeOf(neutral.newtypeBaseType(c.fn()))),
+                        neutral.shapeOf(neutral.newtypeBaseType(c.fn()))));
     }
 
     private Object record(Ast.NewData nd) {
@@ -1320,13 +1113,13 @@ public final class ExampleVerifier {
         // a value's body reaches here already written the second way. Either spelling is the newtype's
         // own neutral form — its inner value — not a field map.
         TypeName built = nd.typeName().denotes();
-        if (isNewtype(built) && nd.spreads().isEmpty() && nd.inits().size() == 1
+        if (neutral.isNewtype(built) && nd.spreads().isEmpty() && nd.inits().size() == 1
                 && nd.inits().get(0).name().equals("value")) {
-            return adjacentlyTagged(nd.typeName().written(),
-                    shaped(raw(nd.inits().get(0).value(), shapeOf(newtypeBaseType(built))),
-                            shapeOf(newtypeBaseType(built))));
+            return neutral.adjacentlyTagged(nd.typeName().written(),
+                    neutral.shaped(raw(nd.inits().get(0).value(), neutral.shapeOf(neutral.newtypeBaseType(built))),
+                            neutral.shapeOf(neutral.newtypeBaseType(built))));
         }
-        Map<String, Ast.TypeRef> declared = fieldTypes(nd.typeName().denotes());
+        Map<String, Ast.TypeRef> declared = neutral.fieldTypes(nd.typeName().denotes());
         Map<String, Object> map = new LinkedHashMap<>();
         // `...base` copies the fields of a value, and the fields written after it replace what it
         // brought. Only a value can be spread here: a row has no bindings to read from.
@@ -1347,8 +1140,8 @@ public final class ExampleVerifier {
             }
         }
         for (Ast.FieldInit fi : nd.inits()) {
-            Object v = shaped(raw(fi.value(), shapeOf(declared.get(fi.name()))),
-                    shapeOf(declared.get(fi.name())));
+            Object v = neutral.shaped(raw(fi.value(), neutral.shapeOf(declared.get(fi.name()))),
+                    neutral.shapeOf(declared.get(fi.name())));
             // `None` on a `T?` field yields a null; leave the key out so the optional decoder reads
             // it as absent (spec 8, absent -> None), the same neutral form as omitting the field.
             if (v == null) {
@@ -1356,140 +1149,8 @@ public final class ExampleVerifier {
             }
             map.put(fi.name(), v);
         }
-        tagged(nd.typeName().denotes(), map);
+        neutral.tagged(nd.typeName().denotes(), map);
         return map;
-    }
-
-    /**
-     * The declared type of a field, used only to shape the written value (a map's entry pairs, a
-     * set's list). The {@code TypeRef} comes from the module that declares the data, and it says what
-     * it denotes — resolved where it was written, so naming a type this module never imported is not
-     * a question asked here at all (issue #110 was that question being asked, and answered with the
-     * declaring file's position).
-     */
-    private Type shapeOf(Ast.TypeRef declaredType) {
-        return declaredType == null ? null : declaredType.denotes();
-    }
-
-    /**
-     * Writes the discriminator a sum's decoder reads, when the constructed data is a case of one.
-     * A fixture names the case it means — `予算枠 = 未定予算`, or a whole `プロジェクト依頼 { … }` where
-     * the parameter is the sum `依頼` — the same way the domain writes a construction, while the
-     * decoder that reads it wants a tag on a key. Where the case is read as itself rather than
-     * through its sum, the tag is a field the decoder does not look at.
-     *
-     * <p>A field the fixture wrote itself is never replaced. A case whose own field is named like its
-     * sum's discriminator is already ambiguous at the boundary — the case encoder and the sum encoder
-     * both claim that key — and overwriting here would hide that behind an example that passes while
-     * decoding something the author did not write. Leaving the written value in place makes the row
-     * fail on the tag it cannot match, which is the honest outcome.
-     */
-    private void tagged(String written, Map<String, Object> map) {
-        TypeName caseName = symbols.resolve(written);
-        if (caseName != null) {
-            tagged(caseName, map);
-        }
-    }
-
-    private void tagged(TypeName caseName, Map<String, Object> map) {
-        for (Ast.Def def : symbols.visible()) {
-            if (!(def instanceof Ast.SumData sum) || sum.decoder().isEmpty()) {
-                continue;
-            }
-            for (Ast.Variant variant : sum.decoder().get().variants()) {
-                if (caseName.equals(variant.caseType().denotes())) {
-                    map.putIfAbsent(sum.decoder().get().key(), variant.tag());
-                    return;
-                }
-            }
-        }
-    }
-
-    /** A data's fields by name, following the `...includes` it composes in (spec §data). */
-    private Map<String, Ast.TypeRef> fieldTypes(TypeName typeName) {
-        Map<String, Ast.TypeRef> out = new LinkedHashMap<>();
-        if (symbols.get(typeName) instanceof Ast.Data d) {
-            for (Ast.Name inc : d.includes()) {
-                out.putAll(fieldTypes(inc.denotes()));
-            }
-            for (Ast.Field f : d.fields()) {
-                // an example builds its input through a decoder, so a field with no external
-                // representation is not one it can state; the data declaration refused it already
-                if (f.type() instanceof Ast.TypeRef ref) {
-                    out.put(f.name(), ref);
-                }
-            }
-        }
-        return out;
-    }
-
-    /** Gives a fixture value the neutral shape its declared type decodes from. A {@code Map} is
-     * written as a list of {@code (key, value)} pairs — Elm's {@code Dict.fromList}, and the same
-     * list literal a {@code Set} takes — while the decoder wants a map, so the pairs are collected
-     * here. Everything else passes through; a {@code List} written as a list stays one, since the
-     * declared type, not the literal, decides. Read on the checked type, so it applies wherever a
-     * value is decoded: a field of a record fixture, and a behavior's argument or output. */
-    private Object shaped(Object v, Type type) {
-        if (type == null || v == null) {
-            return v;
-        }
-        if (type instanceof Type.MapOf map && v instanceof List<?> entries) {
-            Map<Object, Object> m = new LinkedHashMap<>();
-            for (Object entry : entries) {
-                if (!(entry instanceof List<?> pair) || pair.size() != 2) {
-                    throw new FixtureException("a `Map` fixture is a list of (key, value) pairs,"
-                            + " e.g. [ (\"apple\", 3) ]");
-                }
-                m.put(pair.get(0), shaped(pair.get(1), map.value()));
-            }
-            return m;
-        }
-        if (v instanceof List<?> elements) {
-            Type element = switch (type) {
-                case Type.ListOf l -> l.element();
-                case Type.SetOf s -> s.element();
-                default -> null;
-            };
-            if (element != null) {
-                List<Object> out = new ArrayList<>(elements.size());
-                for (Object e : elements) {
-                    out.add(shaped(e, element));
-                }
-                return out;
-            }
-        }
-        return v;
-    }
-
-    private boolean isNewtype(String name) {
-        return symbols.declaration(name) instanceof Ast.Data d && d.newtype();
-    }
-
-    /** As above, for a name that has already been resolved — an imported value's body names its own
-     * module's types, which the module reading the row need not have imported. */
-    private boolean isNewtype(TypeName name) {
-        return name != null && symbols.get(name) instanceof Ast.Data d && d.newtype();
-    }
-
-    private Ast.TypeRef newtypeBaseType(TypeName name) {
-        return name != null && symbols.get(name) instanceof Ast.Data d && d.newtype()
-                && d.fields().size() == 1 && d.fields().get(0).type() instanceof Ast.TypeRef base
-                ? base : null;
-    }
-
-    /** The type a newtype wraps ({@code Date} for {@code data 貸出日 = Date}), or null. */
-    private String newtypeBase(String name) {
-        Ast.TypeRef base = newtypeBaseType(name);
-        return base == null ? null : base.name();
-    }
-
-    /** The written form of what a newtype wraps, kept whole so a generic base
-     * ({@code data 在庫 = Map<商品ID, Int>}) keeps its type arguments. */
-    private Ast.TypeRef newtypeBaseType(String name) {
-        return symbols.declaration(name) instanceof Ast.Data d && d.newtype() && d.fields().size() == 1
-                && d.fields().get(0).type() instanceof Ast.TypeRef base
-                ? base
-                : null;
     }
 
     private static Object negate(Object v) {
@@ -1527,7 +1188,7 @@ public final class ExampleVerifier {
     // --- decode a raw value into the parameter/expected type ----------------------------------
 
     private Object decode(Type type, Object rawValue) {
-        Object raw = shaped(rawValue, type);
+        Object raw = neutral.shaped(rawValue, type);
         Decoder<Object, ?> decoder = decoderFor(type);
         Result<?> result;
         try {
@@ -1721,24 +1382,9 @@ public final class ExampleVerifier {
         return Character.toUpperCase(name.charAt(0)) + name.substring(1);
     }
 
-    /** A fixture that cannot be built (an unsupported form, or an invariant a fixture value breaks). */
-    private static final class FixtureException extends RuntimeException {
-        FixtureException(String message) {
-            super(message);
-        }
-    }
-
     /** The behavior aborted while evaluating a row (e.g. an invariant violation) — a failed example. */
     private static final class AbortException extends RuntimeException {
         AbortException(String message) {
-            super(message);
-        }
-    }
-
-    /** Evaluating a row did not finish within the budget — a `partial` recursion it reaches may not
-     * terminate. Reported as its own diagnostic rather than hanging the compiler. */
-    private static final class NonTerminationException extends RuntimeException {
-        NonTerminationException(String message) {
             super(message);
         }
     }
